makes amendments as per review #6, creates barrel file and utilities.js for reusable hooks/imports. Removes react-google-button
