+++ src/components/auth/Register.jsx
@@ -1,35 +1,29 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import { UserAuth } from "../../context/AuthContext";
-import { Link, useNavigate } from "react-router-dom";
-import GoogleButton from "react-google-button";
+import { Link } from "react-router-dom";
+import { UseRedirectToDashboard } from "./utilities"
+import { RedirectToDashboardOnUser } from "./utilities";
+
 
 const Register = () => {
-  const navigate = useNavigate();
   const { googleSignIn, createUser, user } = UserAuth();
+  RedirectToDashboardOnUser(user);
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [error, setError] = useState("");
 
-  const redirectToDashboard = () => {
-    navigate("/dashboard");
-  };
 
   const handleSubmit = async (e) => {
     e.preventDefault();
     setError("");
     try {
       await createUser(email, password);
-      redirectToDashboard();
+      UseRedirectToDashboard();
     } catch (e) {
       setError(e.message);
     }
   };
 
-  useEffect(() => {
-    if (user != null) {
-      navigate("/dashboard");
-    }
-  }, [user, navigate]);
 
 
   return (
@@ -60,7 +54,7 @@ const Register = () => {
         <button type="submit">Sign Up</button>
       </form>
 
-      <GoogleButton onClick={googleSignIn} />
+      <button onClick={googleSignIn}>Sign Up With Google</button>
 
       {error && <p>{error}</p>}
     </div>
