+++ src/pages/index.js
@@ -1,0 +1,3 @@
+export { default as Home } from './Home';
+export { default as About } from './About';
+export { default as Contact } from './Contact';
